Fix Order.quantity fractional support + book-ticker callback timestamp (#13)
* fix: Order.quantity int→Decimal, book-ticker ts_ns, version dynamic

[HIGH] order.py: quantity field changed from int to Decimal (mirrors
BinanceFuturePosition.quantity design). Imports Decimal. Wire param
str(order.quantity) is unchanged (str(Decimal) is correct).

[HIGH] client.py:395-396: remove int() coercion on deal_quantity and
order_quantity in success path; preserve original Decimal/numeric type.

[HIGH] quote.py: _dispatch_book_ticker global callback was passing mid
price as third arg instead of ns timestamp. Now reads T (transaction
time ms) and converts to ns, matching _dispatch_mark_price convention.

[MED] __init__.py: replace literal "0.5.0" with importlib.metadata
dynamic version (fallback "0.5.1" matches pyproject.toml).

README: add `from decimal import Decimal` import, update quantity
example to Decimal("0.001") (fractional BTC order).

tests/test_quote.py: update global_tick book-ticker assertion to verify
third element is ts_ns (not a price duplicate).


* [fix/review-20260531] fix(trade): order_quantity int→Decimal + fractional test

sdk#13 review triage 兩點（MERGE-WITH-CHANGES）:
(a) BinanceTradeStatus.order_quantity: Optional[int] → Optional[Decimal]
    修 client.py:396 mypy [arg-type]（order.quantity 是 Decimal）。
(b) 補 test_place_order_via_fractional_quantity_preserves_precision：
    驗 Decimal("0.001") 經 place_order_via 序列化為 "0.001" 不丟精度。

驗證（core venv, PYTHONPATH=wt, --no-incremental / no:cacheprovider）:
- mypy: client.py:396 [arg-type] 消失（8→7 errors，剩 7 為既有 TYPE_CHECKING
  name-defined，非本任務範圍）
- pytest: 174 passed（baseline 173 +1 新 test）


* [fix/review-20260531] feat(client): set_leverage / set_margin_type vendor methods

Step A 第二小步 — Binance 純 vendor futures setup（Shioaji 無對應概念）:
- set_leverage(symbol, leverage) → POST /fapi/v1/leverage，回傳 broker 確認值
- set_margin_type(symbol, margin_type) → POST /fapi/v1/marginType；
  -4046 'No need to change' 視為冪等成功（回 'already-set'）
- 錯誤遵循既有慣例用 BinanceAccountError（不新增 error 類）
- _FakeRest 補 post()；+7 tests（成功/REST錯/-4046/非法值/非正數）

驗證（core venv, PYTHONPATH=wt, --no-incremental）:
- mypy client.py+trade.py: Success no issues
- pytest: 181 passed（174 +7）


---------

+++ binance_shioaji_sdk/trade.py
@@ -6,6 +6,7 @@ Trade = contract + order + status, where status is OrderStatusInfo equiv
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 from enum import StrEnum
 from typing import Optional
 
@@ -37,7 +38,7 @@ class BinanceTradeStatus:
     status_code: str
     order_datetime: str                       # ISO 8601 UTC
     deal_quantity: Optional[int] = None       # cumulative filled qty (Binance: executedQty)
-    order_quantity: Optional[int] = None      # original submitted qty (Binance: origQty)
+    order_quantity: Optional[Decimal] = None  # original submitted qty (Binance: origQty)
     cancel_quantity: Optional[int] = None     # qty remaining after cancellation
     modified_price: float = 0.0               # weighted avg fill price (0.0 if no fill)
     msg: str = ""                             # broker message (error reason on failure)
